requirements.txt
# SMPL-X Contact Prediction - Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0

# SMPL-X model
smplx>=0.1.28

# Visualization
matplotlib>=3.7.0
opencv-python>=4.7.0

# Data processing
pyyaml>=6.0
tqdm>=4.65.0

# Mesh processing (optional, for normal computation)
trimesh>=3.21.0

# Development tools (optional)
tensorboard>=2.13.0
jupyter>=1.0.0
